LQI/requirements.txt
pytorch
tqdm
torchsummary